Restructured files for migration to lazy
