+++ init.lua
@@ -1,23 +1,23 @@
 -- Options
-require('core.options')
-require('core.keymaps')
-require('core.plugins')
+require('options')
+require('keymaps')
+require('lazy')
 
 -- Plugins
-require('plugin-settings.colorscheme')
-require('plugin-settings.lsp')
-require('plugin-settings.cmp')
-require('plugin-settings.telescope')
-require('plugin-settings.project')
-require('plugin-settings.treesitter')
-require('plugin-settings.autopairs')
-require('plugin-settings.comment')
-require('plugin-settings.gitsigns')
-require('plugin-settings.nvim-tree')
-require('plugin-settings.toggleterm')
-require('plugin-settings.lualine')
-require('plugin-settings.alpha')
-require('plugin-settings.vimtex')
+require('plugins.colorscheme')
+require('plugins.lsp')
+require('plugins.cmp')
+require('plugins.telescope')
+require('plugins.project')
+require('plugins.treesitter')
+require('plugins.autopairs')
+require('plugins.comment')
+require('plugins.gitsigns')
+require('plugins.nvim-tree')
+require('plugins.toggleterm')
+require('plugins.lualine')
+require('plugins.alpha')
+require('plugins.vimtex')
 
 -- Snippets
-require('plugin-settings.luasnip')
+require('plugins.luasnip')
