--- lua/plugin-settings/lsp/init.lua
@@ -1,8 +1,0 @@
-local status_ok, _ = pcall(require, "lspconfig")
-if not status_ok then
-	return
-end
-
-require("plugin-settings.lsp.mason")
-require("plugin-settings.lsp.handlers").setup()
-require("plugin-settings.lsp.null-ls")
